emergent_testing: the catching runner belongs in a .mjs proof
Review found try/catch in proof.f.mjs — outside the language the file
claims to be written in, and the same defect the repository has been
removing rather than excusing. The irony is exact: the proof exists
because FunctionalScript cannot catch, and it was written with a catch.

`returnedTreeThrows` moves to catch.proof.mjs, where a host runner
belongs. Its mock accumulates written output in the runner's own state
rather than pushing into a captured array, which was the second finding
and is the better shape anyway — the proof now reads like a virtual run.

A second proof comes with the move: a readable returned tree is still
walked, so the guard did not quietly replace recursion with a refusal to
recurse.

The changelog entry is marked breaking. Adding `Catch` to `NodeOp` makes
a handler mandatory for any exhaustive `NodeOperationMap` — which this
change proved by breaking the Node interpreter's typecheck and the mock
map in `reporterWriteFailure` before supplying both.
