+++ changelog/unreleased/1758.md
@@ -1,4 +1,4 @@
-- `emergent_testing`: a proof whose return value cannot be enumerated — an
-  enumerable getter or a proxy trap that throws — is now reported as that
-  proof's failure instead of taking the whole run down with it. `effects`: adds
-  the `catch` operation this needs, beside `sandbox`
+- **BREAKING CHANGES:** `effects`: the new `catch` operation joins `NodeOp`, so
+  an exhaustive `NodeOperationMap` must supply a handler for it
+- `emergent_testing`: a proof whose return value cannot be enumerated fails that
+  proof instead of the whole run
